增加顶部状态栏 增加侧边按钮栏 左侧弹出栏 修复攻击时打塔楼数据错误的问题 实现护盾获得动画飞向定点位置的功能

+++ Assets/Scripts/Data/MessageResponseData.cs
@@ -6,7 +6,7 @@ public class MessageResponseData {
     public long uid;                     // 用户ID
     public long toid;                    // 暂时未用到
     public int action;              // 1 表示摧毁建筑 2 表示偷取
-    public int result;              // 暂时未用到
+    public long result;              // 暂时未用到
     public string time;                // 时间
     public string name;                // 昵称
     public string headImg;              // 头像url
